+++ wagtailtinymce/static/wagtailtinymce/js/tinymce-editor.js
@@ -1,34 +1,9 @@
-/*
-Copyright (c) 2016, Isotoma Limited
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
-    * Neither the name of the Isotoma Limited nor the
-      names of its contributors may be used to endorse or promote products
-      derived from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL ISOTOMA LIMITED BE LIABLE FOR ANY
-DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-
 'use strict';
-
-var mcePlugins = ['hr', 'code', 'fullscreen', 'noneditable', 'paste', 'table'],
-    mceTools = ['inserttable'],
+var mcePlugins = ['code', 'lists', 'nonbreaking', 'visualchars', 'wordcount', 'codeeditor', 'media'],
+    mceTools = [],
+    // mceToolbar = 'code undo bold styleselect',
+    // mceToolbar = 'undo redo | bold italic underline strikethrough | code' ,
+    // mceMenubar = ['insert'],
     mceExternalPlugins = {};
 
 function registerMCEPlugin(name, path, language) {
@@ -48,12 +23,21 @@ function registerMCETool(name) {
 
 function makeTinyMCEEditable(id, kwargs) {
 
-    kwargs = kwargs || {};
+    kwargs = kwargs || {valid_elements : "span[*]"};
     $.extend(kwargs, {
         selector: '#' + id.toString(),
+        height: 400,
         plugins: mcePlugins,
         tools: mceTools,
         external_plugins: mceExternalPlugins,
+        branding: false,
+        media_live_embeds: true,
+        codeeditor_themes_pack: "merbivore",
+        end_container_on_empty_block: true,
+        extended_valid_elements : 'video[autoplay|muted|loop|playsinline|class|width|height|poster|controls]',
+        custom_elements: 'picture',
+        valid_children: 'picture[source|img]',
+        relative_urls: false,
         setup: function (editor) {
             editor.on('change', function () {
                 editor.save();
@@ -61,5 +45,7 @@ function makeTinyMCEEditable(id, kwargs) {
         }
     });
 
-    tinymce.init(kwargs);
+    setTimeout(function () {
+        tinymce.init(kwargs);
+    }, 1);
 }