+++ wagtailtinymce/static/wagtailtinymce/js/tinymce-plugins/wagtailembeds.js
@@ -1,30 +1,3 @@
-/*
-Copyright (c) 2016, Isotoma Limited
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
-    * Neither the name of the Isotoma Limited nor the
-      names of its contributors may be used to endorse or promote products
-      derived from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL ISOTOMA LIMITED BE LIABLE FOR ANY
-DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-
 (function() {
     'use strict';
 
@@ -76,6 +49,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                 ModalWorkflow({
                     url: url,
                     urlParams: urlParams,
+                    onload: EMBED_CHOOSER_MODAL_ONLOAD_HANDLERS,
                     responses: {
                         embedChosen: function(embedData) {
                             var elem = $(embedData).get(0);
@@ -89,20 +63,20 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                 });
             }
 
-            editor.addButton('wagtailembed', {
+            editor.ui.registry.addButton('wagtailembed', {
                 icon: 'media',
-                tooltip: 'Insert/edit embed',
-                onclick: showDialog,
-                stateSelector: '[data-embedtype=media]'
-            });
+                text: 'Embed',
+                onAction: showDialog,
+                stateSelector:  '[data-embedtype=media]'
+            })
 
-            editor.addMenuItem('wagtailembed', {
+            editor.ui.registry.addMenuItem('wagtailembed', {
                 icon: 'media',
-                text: 'Insert/edit embed',
-                onclick: showDialog,
+                text: 'Embed',
+                onAction: showDialog,
                 context: 'insert',
                 prependToContext: true
-            });
+            })
 
             editor.addCommand('mceWagtailEmbed', showDialog);