+++ wagtailtinymce/static/wagtailtinymce/js/tinymce-plugins/wagtailimage.js
@@ -1,30 +1,3 @@
-/*
-Copyright (c) 2016, Isotoma Limited
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
-    * Neither the name of the Isotoma Limited nor the
-      names of its contributors may be used to endorse or promote products
-      derived from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL ISOTOMA LIMITED BE LIABLE FOR ANY
-DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-
 (function() {
     'use strict';
 
@@ -33,7 +6,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
             /* stop editing and resizing of embedded image content */
             function fixContent() {
-                $(editor.getBody()).find('[data-embedtype=image]').each(function () {
+                $(editor.getBody()).find('picture').each(function () {
                     $(this).attr('contenteditable', false).attr('data-mce-contenteditable', 'false').find('div,table,img').attr('data-mce-resize', 'false');
                 });
             }
@@ -74,6 +47,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                     // select and insert after target
                     insertElement = function(elem) {
                         $(elem).insertBefore($targetNode);
+                        // editor.insertContent(elem);
                         mceSelection.select(elem);
                     };
                 }
@@ -81,11 +55,11 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                 ModalWorkflow({
                     url: url,
                     urlParams: urlParams,
+                    onload: IMAGE_CHOOSER_MODAL_ONLOAD_HANDLERS,
                     responses: {
                         imageChosen: function(imageData) {
                             var elem = $(imageData.html).get(0);
                             editor.undoManager.transact(function() {
-                                editor.focus();
                                 insertElement(elem);
                                 fixContent();
                             });
@@ -94,20 +68,20 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                 });
             }
 
-            editor.addButton('wagtailimage', {
+            editor.ui.registry.addButton('wagtailimage', {
                 icon: 'image',
-                tooltip: 'Insert/edit image',
-                onclick: showDialog,
+                text: 'Image',
+                onAction: showDialog,
                 stateSelector: '[data-embedtype=image]'
-            });
+            })
 
-            editor.addMenuItem('wagtailimage', {
+            editor.ui.registry.addMenuItem('wagtailimage', {
                 icon: 'image',
-                text: 'Insert/edit image',
-                onclick: showDialog,
+                text: 'Image',
+                onAction: showDialog,
                 context: 'insert',
                 prependToContext: true
-            });
+            })
 
             editor.addCommand('mceWagtailImage', showDialog);