+++ wagtailtinymce/static/wagtailtinymce/js/tinymce-plugins/wagtaillink.js
@@ -1,30 +1,3 @@
-/*
-Copyright (c) 2016, Isotoma Limited
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in the
-      documentation and/or other materials provided with the distribution.
-    * Neither the name of the Isotoma Limited nor the
-      names of its contributors may be used to endorse or promote products
-      derived from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL ISOTOMA LIMITED BE LIABLE FOR ANY
-DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-
 (function() {
     'use strict';
 
@@ -61,7 +34,8 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                 url = window.chooserUrls.pageChooser;
                 urlParams = {
                     'allow_external_link': true,
-                    'allow_email_link': true
+                    'allow_email_link': true,
+                    'allow_anchor_link': true,
                 };
 
                 mceSelection = editor.selection;
@@ -80,6 +54,11 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                     else if (href.startsWith('mailto:')) {
                         url = window.chooserUrls.emailLinkChooser;
                         href = href.replace('mailto:', '');
+                        urlParams['link_url'] = href;
+                    }
+                    else if (href.startsWith('#')) {
+                        url = window.chooserUrls.anchorLinkChooser;
+                        href = href.replace('#', '');
                         urlParams['link_url'] = href;
                     }
                     else if (!linkType) {
@@ -125,6 +104,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                 ModalWorkflow({
                     url: url,
                     urlParams: urlParams,
+                    onload: PAGE_CHOOSER_MODAL_ONLOAD_HANDLERS,
                     responses: {
                         pageChosen: function(pageData) {
                             editor.undoManager.transact(function() {
@@ -136,30 +116,28 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
                 });
             }
 
-            editor.addButton('link', {
+            editor.ui.registry.addButton('wagtaillink', {
                 icon: 'link',
-                tooltip: 'Insert/edit link',
-                shortcut: 'Meta+K',
-                onclick: showDialog,
-                stateSelector: 'a[data-linktype=page],a[href]:not([data-linktype])'
-            });
+                text: 'Insert link',
+                onAction: showDialog,
+                stateSelector:  'a[data-linktype=page],a[href]:not([data-linktype])'
+            })
 
-            editor.addButton('unlink', {
-                icon: 'unlink',
-                tooltip: 'Remove link',
-                cmd: 'unlink',
-                stateSelector: 'a[data-linktype=page],a[href]'
-            });
+            // editor.ui.registry.addButton('unlink', {
+            //     icon: 'unlink',
+            //     text: 'Remove link',
+            //     onAction: showDialog,
+            //     stateSelector:  'a[data-linktype=page],a[href]'
+            // })
 
-            editor.addMenuItem('link', {
+            editor.ui.registry.addMenuItem('wagtaillink', {
                 icon: 'link',
-                text: 'Insert/edit link',
-                shortcut: 'Meta+K',
-                onclick: showDialog,
-                stateSelector: 'a[data-linktype=page],a[href]:not([data-linktype])',
+                text: 'Remove link',
+                onAction: showDialog,
+                 stateSelector: 'a[data-linktype=page],a[href]:not([data-linktype])',
                 context: 'insert',
                 prependToContext: true
-            });
+            })
 
             editor.addShortcut('Meta+K', '', showDialog);
             editor.addCommand('mceLink', showDialog);